Changed BSD license to MIT X11.
This license change shouldn't affect most people as the licenses are pretty similar.
Commercial use is and will always be permitted...

This license change was made to make it easier for GSoC contributers to contribute to dotGit.
It will also make it easier for the Mono project to use dotGit because they're using MIT X11 as well.